LRCI-1174 Rename update jira ticket utility class
